Support for more specific toolchain versions
This adds support for more specific versions of a given prebuilt
toolchain. In addition to $(HOST_PREBUILT_TAG) (like darwin-ppc or
linux-x86) a new $(HOST_PREBUILT_EXTRA_TAG) is introduced containing (so
far) the number of address bits (32/64) for the build host or the same
value as $(HOST_PREBUILT_TAG) if this could not be determined.

Subsequently everything in the building process that needs to determine
the full path for a prebuilt tool should check with
$(HOST_PREBUILT_EXTRA_TAG) first and then fall back to
$(HOST_PREBUILT_TAG).

The advantage for the user is that she can place a 64-bit or a 32-bit
version of, say, gcc and binutils next to the default prebuilts without
messing around with the directory structure, repo manifest and having a
working fallback to the default binaries.

Long story short: To make Jelly Bean build on 32-bit hosts again,
something like
 <project
  name="yath/android_prebuilts_gcc_linux-x86-32_arm_arm-linux-androideabi-4.6"
  path="prebuilts/gcc/linux-x86-32/arm/arm-linux-androideabi-4.6" />
can be added to local_manifest.xml. make will pick the 32-bit versions
of gcc and binutils instead of trying to run the new 64-bit ones.

Change-Id: I89ed0f00eda1a0bebbc35f39af4bbf0e97f6fade

Conflicts:

	core/tasks/kernel.mk